Have a presenter per downstream system…
These are quite similar, so for now – we have a
DownStream presenter that does the bulk of the
work.
