+++ spec/pacts/publishing_api-content_store.json
@@ -45,7 +45,7 @@
           ],
           "links": {
             "organisations": [
-              "bdeef46d-74fd-40f8-b3e8-fe12ea472f6b"
+              "35d1de40-0e0f-4475-9559-c03f0bcfb699"
             ]
           },
           "transmitted_at": "2000000000000000000"
@@ -99,7 +99,7 @@
           ],
           "links": {
             "organisations": [
-              "211ef118-77b3-4ae8-9a6c-b1e8a35651df"
+              "9756ccde-0b70-4e2d-b821-db137131876e"
             ]
           },
           "transmitted_at": "2000000000000000000"
